requirements.txt
python-telegram-bot==22.0
SQLAlchemy==2.0.19
jira==3.5.1
networkx==3.1
numpy==1.24.3
Pillow==10.0.0
python-dotenv==1.0.0